chore(deps): bump alloy to 0.4.2 (#9000)
* chore: bump alloy-core deps + revm

* bump alloy to 0.4.0

* bump revm-inspectors + type casting to u128

* fix

* fix

* fix

* bump foundry-fork-db

* bump alloy

* fix

* gas related field to u64

* fmt

* change gas fields types to u64 in DepositTx, TxEssentials

+++ crates/cheatcodes/src/inspector.rs
@@ -677,11 +677,7 @@ impl Cheatcodes {
                             value: Some(input.value()),
                             input: TransactionInput::new(input.init_code()),
                             nonce: Some(account.info.nonce),
-                            gas: if is_fixed_gas_limit {
-                                Some(input.gas_limit() as u128)
-                            } else {
-                                None
-                            },
+                            gas: if is_fixed_gas_limit { Some(input.gas_limit()) } else { None },
                             ..Default::default()
                         }
                         .into(),
@@ -1016,11 +1012,7 @@ impl Cheatcodes {
                             value: call.transfer_value(),
                             input: TransactionInput::new(call.input.clone()),
                             nonce: Some(account.info.nonce),
-                            gas: if is_fixed_gas_limit {
-                                Some(call.gas_limit as u128)
-                            } else {
-                                None
-                            },
+                            gas: if is_fixed_gas_limit { Some(call.gas_limit) } else { None },
                             ..Default::default()
                         }
                         .into(),
