+++ crates/evm/core/src/utils.rs
@@ -84,7 +84,7 @@ pub fn get_function<'a>(
 /// Configures the env for the transaction
 pub fn configure_tx_env(env: &mut revm::primitives::Env, tx: &Transaction) {
     env.tx.caller = tx.from;
-    env.tx.gas_limit = tx.gas as u64;
+    env.tx.gas_limit = tx.gas;
     env.tx.gas_price = U256::from(tx.gas_price.unwrap_or_default());
     env.tx.gas_priority_fee = tx.max_priority_fee_per_gas.map(U256::from);
     env.tx.nonce = Some(tx.nonce);
